requirements.txt
aws-cdk-lib==2.1005.0
cdk-nag==2.28.164
constructs>=10.0.0,<11.0.0
cfn-lint
cdklabs.cdk-validator-cfnguard
requests
pandas
pyarrow
tabulate
aws-pdk
asyncio
-e .